pymemorybox/pyproject.toml
[project]
name = "pymemorybox"
dynamic = ["version"]
authors = [
  { name="Guillaume BITON", email="guillaume@gbweb.fr" },
]
description = "An app/service to reveal one picture a day (and print it)."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Development Status :: 2 - Pre-Alpha",
    "Framework :: Flask",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Graphics :: Viewers",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content"
]
dependencies = [
  "click",
  "flask",
  "Flask-SQLAlchemy",
  "flask-login",
  "pybluez @ git+https://github.com/pybluez/pybluez@master#egg=pybluez",
  "peripage",
  "BeautifulSoup4",
  "requests",
  "toml",
  "gunicorn",
  "flask-socketio",
  "eventlet",
  "websocket-client",
  "python-socketio",
  'uuid'
]

[project.urls]
Homepage = "https://github.com/gbwebdev/memorybox"
Issues = "https://github.com/gbwebdev/memorybox/issues"

[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project.scripts]
memorybox = "pymemorybox:cli"


[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
pymemorybox = [
    "static/**/*",
    "templates/**/*",
    "templates/*"
]